Datatype change for foreign key time_id

+++ database/tables/user_login/user.sql
@@ -2,5 +2,5 @@ CREATE TABLE IF NOT EXISTS "user_"(
 id SERIAL PRIMARY KEY,
 username TEXT UNIQUE NOT NULL,
 auth_id int REFERENCES auth(id) NOT NULL,
-user_time int REFERENCES "time"(id) NOT NULL
+user_time bigint REFERENCES "time"(id) NOT NULL
 );
